impl NamedFrom for ChunkedArray
